pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["docs/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
exclude = ["^tests/"]

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[project]
name = "ragx"
dynamic = ["version"]
dependencies = [
    "jinja2",
    "jsonschema",
    "deepdiff",
    "numpy",
    "packaging",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pybind11",
    "pytest",
    "pytest-cov",
    "ruff",
    "yamllint",
]
